Fix "Stop indexing Repos" Bug in registry Package
In the File `pkg/registry/api.go` ain Function DockerRegistry.FetchAll()
the Index Progress of the Repositories is stopped if a Repo is found
with an equal Name.
Why? It should skip that repo and not stop the hole Index Progress.

Rename used variable _doBreak_ to _skipRepo_ which does describe the
usage of the Variable better then _doBreak_.